requirements.txt
Django
ipython
remote-pdb
requests
powerline-status
celery
django-celery-results
python-dateutil
flower
psycopg2-binary
python-telegram-bot
cython